refactor(providers): delete legacy providers/github submodule files + paired tests (resolves #1379)
