+++ baselines/maintainability.json
@@ -244,22 +244,8 @@
   ".agents/scripts/noise-study.js": 88.128,
   ".agents/scripts/notify.js": 89.774,
   ".agents/scripts/post-structured-comment.js": 95.152,
-  ".agents/scripts/providers/github.js": 91.728,
-  ".agents/scripts/providers/github/auth.js": 109.556,
-  ".agents/scripts/providers/github/branches.js": 84.374,
-  ".agents/scripts/providers/github/cache-manager.js": 92.041,
-  ".agents/scripts/providers/github/comments.js": 115.955,
-  ".agents/scripts/providers/github/error-classifier.js": 87.335,
-  ".agents/scripts/providers/github/graphql-builder.js": 117.577,
-  ".agents/scripts/providers/github/graphql.js": 146.779,
-  ".agents/scripts/providers/github/http-client.js": 94.588,
-  ".agents/scripts/providers/github/http.js": 171,
-  ".agents/scripts/providers/github/issues.js": 93.637,
-  ".agents/scripts/providers/github/labels.js": 106.765,
+  ".agents/scripts/providers/github.js": 91.412,
   ".agents/scripts/providers/github/projects-v2-graphql.js": 95.347,
-  ".agents/scripts/providers/github/projects.js": 93.266,
-  ".agents/scripts/providers/github/repo.js": 107.958,
-  ".agents/scripts/providers/github/ticket-mapper.js": 106.843,
   ".agents/scripts/quality-preview.js": 91.369,
   ".agents/scripts/quality-watch.js": 0,
   ".agents/scripts/render-manifest.js": 92.206,
@@ -500,7 +486,6 @@
   "tests/lib/planning-state-manager.test.js": 85.084,
   "tests/lib/presentation/manifest-formatter-end-to-end.test.js": 100.773,
   "tests/lib/presentation/manifest-formatter-layout.test.js": 98.735,
-  "tests/lib/providers/github-branches.test.js": 122.358,
   "tests/lib/providers/github-projects-extra.test.js": 106.189,
   "tests/lib/reconciler.test.js": 113.04,
   "tests/lib/recut.test.js": 119.942,
@@ -548,16 +533,8 @@
   "tests/plan-phase-cleanup.test.js": 117.662,
   "tests/post-structured-comment-cli.test.js": 106.937,
   "tests/pre-commit-hook.test.js": 115.365,
-  "tests/provider-factory.test.js": 119.906,
-  "tests/providers-github-cache-manager.test.js": 107.267,
-  "tests/providers-github-error-classifier.test.js": 106.345,
-  "tests/providers-github-graphql-builder.test.js": 121.639,
-  "tests/providers-github-sub-issue-link.test.js": 94.951,
-  "tests/providers-github-subtickets.test.js": 100.138,
-  "tests/providers-github-ticket-mapper.test.js": 106.932,
-  "tests/providers-github-token-memoize.test.js": 107.374,
-  "tests/providers-github.test.js": 97.735,
-  "tests/providers/ticket-cache-maxage.test.js": 101.985,
+  "tests/provider-factory.test.js": 103.85,
+  "tests/providers-github.test.js": 97.608,
   "tests/push-epic-retry.test.js": 98.687,
   "tests/quality-preview.test.js": 92.429,
   "tests/quality-watch.test.js": 107.32,
@@ -566,7 +543,7 @@
   "tests/run-audit-suite-cli.test.js": 106.961,
   "tests/schemas/epic-perf-report.schema.test.js": 104.452,
   "tests/schemas/signal-schemas.test.js": 105.795,
-  "tests/scripts/agents-bootstrap-github-protection.test.js": 98.807,
+  "tests/scripts/agents-bootstrap-github-protection.test.js": 103.888,
   "tests/scripts/epic-deliver-automerge.test.js": 113.837,
   "tests/scripts/epic-deliver-baseline-reconciliation.test.js": 106.786,
   "tests/scripts/epic-deliver-cleanup.test.js": 109.479,
