+++ tests/scripts/agents-bootstrap-github-protection.test.js
@@ -184,121 +184,11 @@ describe('agents-bootstrap-github — ensureMainBranchProtection', () => {
   });
 });
 
-// Provider-layer behaviour test: the additive merge math lives inside
-// `providers/github/branches.setBranchProtection`. We exercise it
-// directly so the bootstrap-orchestrator tests above can stay focused on
-// the wiring contract rather than the merge algorithm.
-describe('providers/github/branches.setBranchProtection — additive merge', () => {
-  it('PUTs the union of existing + supplied contexts when a rule already exists', async () => {
-    const httpCalls = [];
-    const ctx = {
-      owner: 'acme',
-      repo: 'widgets',
-      http: {
-        async rest(endpoint, opts) {
-          httpCalls.push({ endpoint, opts });
-          if (!opts || opts.method !== 'PUT') {
-            // GET path — return an existing rule.
-            return {
-              required_status_checks: {
-                strict: true,
-                contexts: ['existing/check', 'lint'],
-              },
-              enforce_admins: { enabled: true },
-              required_pull_request_reviews: {
-                required_approving_review_count: 1,
-              },
-              restrictions: null,
-            };
-          }
-          return {};
-        },
-      },
-    };
-
-    const { setBranchProtection } = await import(
-      pathToFileURL(
-        path.join(
-          ROOT,
-          '.agents',
-          'scripts',
-          'providers',
-          'github',
-          'branches.js',
-        ),
-      ).href
-    );
-
-    const result = await setBranchProtection(ctx, 'main', {
-      contexts: ['lint', 'test'],
-    });
-
-    assert.equal(result.created, false);
-    assert.deepEqual(result.added, ['test']);
-    assert.deepEqual(result.existing, ['existing/check', 'lint']);
-
-    // Two HTTP calls: GET (read existing) → PUT (write merged).
-    assert.equal(httpCalls.length, 2);
-    const put = httpCalls[1];
-    assert.equal(put.opts.method, 'PUT');
-    assert.deepEqual(put.opts.body.required_status_checks.contexts, [
-      'existing/check',
-      'lint',
-      'test',
-    ]);
-    // Operator-tuned PR-review and admin-enforce settings flow through
-    // unchanged so the bootstrap re-run doesn't clobber them.
-    assert.equal(put.opts.body.enforce_admins, true);
-    assert.equal(
-      put.opts.body.required_pull_request_reviews
-        .required_approving_review_count,
-      1,
-    );
-  });
-
-  it('creates a fresh rule with sensible defaults when none exists', async () => {
-    const httpCalls = [];
-    const ctx = {
-      owner: 'acme',
-      repo: 'widgets',
-      http: {
-        async rest(endpoint, opts) {
-          httpCalls.push({ endpoint, opts });
-          if (!opts || opts.method !== 'PUT') {
-            // GET path — simulate 404 by throwing the canonical message
-            // shape `getBranchProtection` recognises.
-            throw new Error('GET /protection failed (404): Not Found');
-          }
-          return {};
-        },
-      },
-    };
-
-    const { setBranchProtection } = await import(
-      pathToFileURL(
-        path.join(
-          ROOT,
-          '.agents',
-          'scripts',
-          'providers',
-          'github',
-          'branches.js',
-        ),
-      ).href
-    );
-
-    const result = await setBranchProtection(ctx, 'main', {
-      contexts: ['lint', 'test'],
-    });
-
-    assert.equal(result.created, true);
-    assert.deepEqual(result.added, ['lint', 'test']);
-    assert.deepEqual(result.existing, []);
-    const put = httpCalls.find((c) => c.opts?.method === 'PUT');
-    assert.ok(put);
-    assert.equal(put.opts.body.required_status_checks.strict, true);
-    assert.equal(put.opts.body.enforce_admins, false);
-    assert.equal(put.opts.body.required_pull_request_reviews, null);
-    assert.equal(put.opts.body.restrictions, null);
-  });
-});
+// Provider-layer behaviour test: the additive merge math used to live in
+// `providers/github/branches.setBranchProtection`. After Epic #1179 / Story
+// #1363 the algorithm moved onto `GitHubProvider.setBranchProtection`
+// (gh-exec-backed) and is exercised directly by
+// `tests/providers-github.test.js` under the `setBranchProtection()`
+// describe — see the "additively merges contexts when a rule already
+// exists" / "preserves operator overrides" cases. The orchestrator-wiring
+// tests above remain the responsibility of this file.
